requirements.txt
python-telegram-bot[job-queue]
openai
gspread
google-auth
python-dotenv
pydantic
psycopg2-binary
matplotlib
